Allow using multiple configs with same branch
By setting a global configdir variable into the inventory file, we now include
the _env files from that directory. This allows easy creation of multiple
layouts with the same scripts, running the same playbooks but with different
inventory files. You can set it optionally to "" to have the _env.yml files in
the main directory.