Don't install RPM-base mysql connector
Note: we use the mysql-connector-java from the war file, not the system one.
The code determines whether we can use mysql-connector-java-5.1.X or
mysql-connector-java-8.0.X since they have a different class name for the
driver, so we can better ship along the right connector and rely on that.
For CentOS6 we must use the 5.1 version. This version also works on CentOS7.